refactor checks part.1
